Flask - BE/requirements.txt
absl-py==0.12.0
aniso8601==9.0.1
astunparse==1.6.3
autopep8==1.6.0
cachetools==4.2.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.10.0
decorator==5.1.1
dnspython==2.2.1
Flask==2.1.1
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
flask-swagger-ui==3.36.0
flatbuffers==1.12
fpdf==1.7.2
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.32.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.0.1
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.3.4
mysqlclient==2.1.0
networkx==2.6.3
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
Pillow==8.1.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pygame==2.0.1
PyMySQL==1.0.2
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.1
six==1.16.0
SQLAlchemy==1.4.35
Sublist3r==1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tesseract==0.1.3
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.9
validators==0.18.2
Werkzeug==2.1.0
wrapt==1.12.1
zipp==3.4.1
